requirements.txt
# Core Dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Web UI
gradio>=4.0.0

# CLI
rich>=13.0.0
typer>=0.9.0

# Optional: AI Outpainting (Phase 2)
# diffusers>=0.21.0
# transformers>=4.30.0
# torch>=2.0.0

# Dev
pytest>=7.0.0
